v1.0.0.31

+++ SkyWrathSharp/SkyWrathSharp.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Linq;
+using System.Runtime.Remoting.Messaging;
 using Ensage;
 using Ensage.Common;
 using Ensage.Common.Menu;
@@ -198,7 +199,7 @@ namespace SkyWrathSharp
                 || target.HasModifier("modifier_rune_haste")
                 || target.IsMagicImmune()
                 || !Menu.Item("abilities").GetValue<AbilityToggler>().IsEnabled("skywrath_mage_mystic_flare")
-                || target.Health / target.MaximumHealth * 100 < Menu.Item("noCastUlti").GetValue<Slider>().Value) return;
+                || (int)(target.Health / target.MaximumHealth * 100) < Menu.Item("noCastUlti").GetValue<Slider>().Value) return;
 
             if (!target.CanMove() || target.NetworkActivity == NetworkActivity.Idle ||
                 target.UnitState.HasFlag(UnitState.Frozen) || target.UnitState.HasFlag(UnitState.Stunned))
